Attrangs_backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Alizarbukhari <alizar.fsd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["staandard"], version = "^0.34.0"}
fastapi = "^0.115.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.1"
python-multipart = "^0.0.20"
psycopg2-binary = "^2.9.10"
sqlmodel = "^0.0.22"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.10.4"}
email-validator = "^2.2.0"
bcrypt = "3.2.0"
requests = "^2.32.3"

[tool.poetry.scripts]
dev= "app.main:start"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
